Break up `reporter` module into different files
