+++ crates/brioche-core/src/bake/unarchive.rs
@@ -6,6 +6,7 @@ use bstr::BString;
 use crate::{
     blob::BlobHash,
     recipe::{Artifact, Directory, File, Meta, Unarchive, WithMeta},
+    reporter::job::{NewJob, UpdateJob},
     Brioche,
 };
 
@@ -27,7 +28,7 @@ pub async fn bake_unarchive(
 
     tracing::debug!(%blob_hash, archive = ?unarchive.archive, compression = ?unarchive.compression, "starting unarchive");
 
-    let job_id = brioche.reporter.add_job(crate::reporter::NewJob::Unarchive);
+    let job_id = brioche.reporter.add_job(NewJob::Unarchive);
 
     let archive_path = {
         let mut permit = crate::blob::get_save_blob_permit().await?;
@@ -60,10 +61,9 @@ pub async fn bake_unarchive(
                 let estimated_progress =
                     position as f64 / (uncompressed_archive_size as f64).max(1.0);
                 let progress_percent = (estimated_progress * 100.0).min(99.0) as u8;
-                brioche.reporter.update_job(
-                    job_id,
-                    crate::reporter::UpdateJob::Unarchive { progress_percent },
-                );
+                brioche
+                    .reporter
+                    .update_job(job_id, UpdateJob::Unarchive { progress_percent });
 
                 let entry = match archive_entry.header().entry_type() {
                     tar::EntryType::Regular => {
@@ -175,7 +175,7 @@ pub async fn bake_unarchive(
 
         brioche.reporter.update_job(
             job_id,
-            crate::reporter::UpdateJob::Unarchive {
+            UpdateJob::Unarchive {
                 progress_percent: 100,
             },
         );
