+++ crates/brioche-core/src/download.rs
@@ -2,7 +2,10 @@ use anyhow::Context as _;
 use futures::TryStreamExt as _;
 use tokio_util::compat::FuturesAsyncReadCompatExt as _;
 
-use crate::Brioche;
+use crate::{
+    reporter::job::{NewJob, UpdateJob},
+    Brioche,
+};
 
 #[tracing::instrument(skip(brioche, expected_hash))]
 pub async fn download(
@@ -22,7 +25,7 @@ pub async fn download(
 
     let job_id = brioche
         .reporter
-        .add_job(crate::reporter::NewJob::Download { url: url.clone() });
+        .add_job(NewJob::Download { url: url.clone() });
 
     let response = brioche.download_client.get(url.clone()).send().await?;
     let response = response.error_for_status()?;
@@ -52,7 +55,7 @@ pub async fn download(
                 let progress_percent = progress_percent.round().min(99.0) as u8;
                 brioche.reporter.update_job(
                     job_id,
-                    crate::reporter::UpdateJob::Download {
+                    UpdateJob::Download {
                         progress_percent: Some(progress_percent),
                     },
                 );
@@ -73,7 +76,7 @@ pub async fn download(
 
     brioche.reporter.update_job(
         job_id,
-        crate::reporter::UpdateJob::Download {
+        UpdateJob::Download {
             progress_percent: Some(100),
         },
     );
