+++ crates/brioche/src/check.rs
@@ -5,7 +5,7 @@ use brioche_core::project::ProjectHash;
 use brioche_core::project::ProjectLocking;
 use brioche_core::project::ProjectValidation;
 use brioche_core::project::Projects;
-use brioche_core::reporter::ConsoleReporterKind;
+use brioche_core::reporter::console::ConsoleReporterKind;
 use brioche_core::reporter::Reporter;
 use brioche_core::Brioche;
 use clap::Parser;
@@ -25,7 +25,7 @@ pub struct CheckArgs {
 
 pub async fn check(args: CheckArgs) -> anyhow::Result<ExitCode> {
     let (reporter, mut guard) =
-        brioche_core::reporter::start_console_reporter(ConsoleReporterKind::Auto)?;
+        brioche_core::reporter::console::start_console_reporter(ConsoleReporterKind::Auto)?;
 
     let brioche = brioche_core::BriocheBuilder::new(reporter.clone())
         .build()
